Reorganise the routes module
With the removal of rfc003 we now support the following routes

- "/"

- "/peers"

- "/swaps/:swap_id"
- "/swaps/:swap_id/init"
- "/swaps/:swap_id/fund"
- "/swaps/:swap_id/deploy"      # For herc20 swaps
- "/swaps/:swap_id/redeem"
- "/swaps/:swap_id/refund"

- "/swaps/herc20/hbit"
- "/swaps/hbit/herc20"
- "/swaps/herc20/halbit"
- "/swaps/halbit/herc20"

We can model this with the structure of our modules in order to make
the code easier to follow by doing:

- Put the code for the "/" route in `routes.rs`

- Put the code for the "/peers" route in `routes/peers.rs`

- Put the code for the GET requests for swap specific endpoints "/swaps/:id"
  in `routes/swaps.rs`

- Put the code for the POST requests on endpoints staring with
  "/swaps" in the responsible protocol module
