+++ cnd/src/http_api/herc20_halbit.rs
@@ -1,2 +1,51 @@
 pub mod alice;
 pub mod bob;
+
+use crate::{
+    halbit, herc20,
+    http_api::{post::*, problem},
+    network::{Herc20Halbit, Identities},
+    storage::Save,
+    Facade, LocalSwapId,
+};
+use digest::Digest;
+use serde::Deserialize;
+use warp::{http::StatusCode, Rejection, Reply};
+
+pub async fn post_swap(body: serde_json::Value, facade: Facade) -> Result<impl Reply, Rejection> {
+    let body = Body::<Herc20, Halbit>::deserialize(&body)
+        .map_err(anyhow::Error::new)
+        .map_err(problem::from_anyhow)
+        .map_err(warp::reject::custom)?;
+
+    let swap_id = LocalSwapId::default();
+    let reply = warp::reply::reply();
+
+    let swap = body.to_created_swap::<herc20::CreatedSwap, halbit::CreatedSwap>(swap_id);
+    facade
+        .save(swap)
+        .await
+        .map_err(problem::from_anyhow)
+        .map_err(warp::reject::custom)?;
+
+    let identities = Identities {
+        ethereum_identity: Some(body.alpha.identity),
+        lightning_identity: Some(body.beta.identity),
+        bitcoin_identity: None,
+    };
+    let digest = Herc20Halbit::from(body.clone()).digest();
+    let peer = body.peer.into();
+    let role = body.role.0;
+
+    facade
+        .initiate_communication(swap_id, peer, role, digest, identities)
+        .await
+        .map(|_| {
+            warp::reply::with_status(
+                warp::reply::with_header(reply, "Location", format!("/swaps/{}", swap_id)),
+                StatusCode::CREATED,
+            )
+        })
+        .map_err(problem::from_anyhow)
+        .map_err(warp::reject::custom)
+}
